UpstreamManager dependency on TaskManager removed. Separate files for UpstreamManager.
